added project which communicate with oracle database based procedure

+++ src/main/java/com/gl/storedprocedure_springboot_jpa_oracle_demo/repository/UserRepository.java
@@ -16,7 +16,7 @@ import java.util.List;
 
     @Procedure("getNameById")
     Object getNameById(@Param("user_id")Integer userId);*/
-public interface UserRepository extends CrudRepository<User, Long>, UserRepositoryCustom {
+public interface UserRepository extends CrudRepository<User, Integer>, UserRepositoryCustom {
 
 
 }
